bugfix: int overflow

+++ mars/comm/http.cc
@@ -1110,20 +1110,20 @@ Parser::TRecvStatus Parser::Recv(AutoBuffer& _recv_buffer) {
                         _recv_buffer.Move(-(trailerEnd - chunkSizeBegin + 2));
                     }
                 } else {  // no chunk
-                    int contentLength = headfields_.ContentLength();
-                    int appendlen = 0;
+                    int64_t contentLength = headfields_.ContentLength();
+                    int64_t appendlen = 0;
 
-                    if (int(_recv_buffer.Length() + bodyreceiver_->Length()) <= contentLength)
-                        appendlen = int(_recv_buffer.Length());
+                    if (int64_t(_recv_buffer.Length() + bodyreceiver_->Length()) <= contentLength)
+                        appendlen = int64_t(_recv_buffer.Length());
                     else {
-                        xwarn2(TSF"contentLength:%_, body.len:%_, recv len:%_", contentLength, int(bodyreceiver_->Length()), _recv_buffer.Length());
-                        appendlen = contentLength - int(bodyreceiver_->Length());
+                        xwarn2(TSF"contentLength:%_, body.len:%_, recv len:%_", contentLength, int64_t(bodyreceiver_->Length()), _recv_buffer.Length());
+                        appendlen = contentLength - int64_t(bodyreceiver_->Length());
                     }
 
                     bodyreceiver_->AppendData(_recv_buffer.Ptr(), (size_t)appendlen);
                     _recv_buffer.Move(-appendlen);
 
-                    if ((int)bodyreceiver_->Length() == contentLength) {
+                    if ((int64_t)bodyreceiver_->Length() == contentLength) {
                         recvstatus_ = kEnd;
                         bodyreceiver_->EndData();
                         return  recvstatus_;
